re-org code
